fix(cart): resolve route conflict by using storefront cart page
- Removed duplicate cart page at src/app/cart/page.tsx
- Replaced server-side cart page with enhanced client-side implementation
- Uses cart store (localStorage + Context) instead of API fetch
- Adds interactive features: qty controls, remove button, subtotal display

Fixes build error: "You cannot have two parallel pages that resolve to the same path"
