+++ frontend/src/app/(storefront)/cart/page.tsx
@@ -1,43 +1,69 @@
-import CartBadge from '@/components/CartBadge'
+'use client'
+import React from 'react'
+import Link from 'next/link'
+import { useCart } from '@/store/cart'
 
-async function getCart() {
-  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL || ''}/api/cart`, { cache:'no-store' })
-  if (!res.ok) return { items: [], totalCents: 0 }
-  return res.json()
-}
+export default function CartPage() {
+  const { items, subtotal, remove, increase, decrease } = useCart()
 
-export default async function CartPage() {
-  const data = await getCart()
-  const items = data.items || []
-  const total = (data.totalCents ?? 0) / 100
-  const count = items.length
-
+  const empty = !items || items.length === 0
   return (
-    <main className="container mx-auto p-6">
-      <div className="flex items-center justify-between mb-4">
-        <h1 className="text-2xl font-semibold" data-testid="cart-title">
-          Καλάθι {count > 0 ? `(${count})` : ''}
-        </h1>
-        {/* client badge for live updates */}
-        <CartBadge />
-      </div>
+    <main className="min-h-screen bg-gray-50 py-8 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-5xl mx-auto">
+        <h1 className="text-2xl font-bold mb-4">Καλάθι</h1>
 
-      {items.length === 0 ? (
-        <div className="text-sm text-neutral-600">Το καλάθι είναι άδειο.</div>
-      ) : (
-        <div className="space-y-3">
-          {items.map((i:any) => (
-            <div key={i.id} className="cart-row flex items-center justify-between border p-3 rounded">
-              <div>
-                <div className="font-medium">{i.title}</div>
-                <div className="text-xs text-neutral-600">Ποσότητα: {i.qty}</div>
-              </div>
-              <div className="text-sm">{(i.priceCents/100).toFixed(2)} €</div>
+        {empty ? (
+          <div className="bg-white border rounded-xl p-10 text-center">
+            <p className="text-gray-600 mb-4">Το καλάθι σας είναι άδειο.</p>
+            <Link href="/products" className="inline-block bg-emerald-600 text-white px-4 py-2 rounded-lg hover:bg-emerald-700">
+              Συνέχεια αγορών
+            </Link>
+          </div>
+        ) : (
+          <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
+            <div className="lg:col-span-2 bg-white border rounded-xl divide-y">
+              {items.map((it) => (
+                <div key={it.id} className="p-4 flex gap-4 items-center">
+                  <div className="w-20 h-20 bg-gray-100 overflow-hidden rounded">
+                    {it.imageUrl ? (
+                      <img src={it.imageUrl} alt={it.title} className="w-full h-full object-cover"/>
+                    ) : (
+                      <div className="w-full h-full flex items-center justify-center text-gray-400">
+                        📦
+                      </div>
+                    )}
+                  </div>
+                  <div className="flex-1">
+                    <div className="font-semibold leading-tight line-clamp-2">{it.title}</div>
+                    <div className="text-sm text-gray-500">{it.producer || 'Παραγωγός'}</div>
+                    <div className="mt-2 flex items-center gap-3">
+                      <button onClick={() => decrease(it.id)} className="h-8 w-8 rounded border hover:bg-gray-50 flex items-center justify-center">−</button>
+                      <span className="min-w-8 text-center">{it.qty ?? 1}</span>
+                      <button onClick={() => increase(it.id)} className="h-8 w-8 rounded border hover:bg-gray-50 flex items-center justify-center">+</button>
+                      <button onClick={() => remove(it.id)} className="ml-4 text-sm text-red-600 hover:underline">Αφαίρεση</button>
+                    </div>
+                  </div>
+                  <div className="font-semibold whitespace-nowrap">{it.priceFormatted ?? '—'}</div>
+                </div>
+              ))}
             </div>
-          ))}
-          <div className="pt-2 text-right font-semibold">Σύνολο: {total.toFixed(2)} €</div>
-        </div>
-      )}
+
+            <aside className="bg-white border rounded-xl p-6 h-fit">
+              <div className="flex items-center justify-between">
+                <span className="text-gray-600">Υποσύνολο</span>
+                <span className="text-lg font-bold">{subtotal?.formatted ?? '—'}</span>
+              </div>
+              <p className="text-xs text-gray-500 mt-2">Οι τελικές χρεώσεις (μεταφορικά/ΦΠΑ) υπολογίζονται στο checkout.</p>
+              <Link href="/checkout" className="mt-4 w-full inline-flex justify-center bg-emerald-600 text-white px-4 py-2 rounded-lg hover:bg-emerald-700">
+                Συνέχεια στο checkout
+              </Link>
+              <Link href="/products" className="mt-2 w-full inline-flex justify-center border border-gray-300 text-gray-700 px-4 py-2 rounded-lg hover:bg-gray-50">
+                Συνέχεια αγορών
+              </Link>
+            </aside>
+          </div>
+        )}
+      </div>
     </main>
   )
 }
